vector-store: extract column_vec mod from table
